backend/pyproject.toml
[tool.poetry]
name = "smart-notes-backend"
version = "0.1.0"
description = "Smart Notes backend: FastAPI, LLM agent, faster-whisper STT, hybrid search"
authors = ["Alex Tikhonov <ascorblack@gmail.com>"]
readme = "README.md"
packages = [{ include = "app", from = "." }]

[[tool.poetry.source]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
priority = "explicit"

[tool.poetry.dependencies]
python = "^3.12"
torch = { version = ">=2.0,<3", source = "pytorch-cpu" }
fastapi = "==0.115.5"
python-multipart = "==0.0.12"
uvicorn = { version = "==0.32.1", extras = ["standard"] }
sqlalchemy = { version = "==2.0.36", extras = ["asyncio"] }
asyncpg = "==0.30.0"
alembic = "==1.14.0"
pydantic-settings = "==2.6.1"
email-validator = "==2.2.0"
python-jose = { version = "==3.3.0", extras = ["cryptography"] }
bcrypt = "==4.2.1"
httpx = "==0.28.1"
faster-whisper = "==1.0.3"
requests = ">=2.28.0"
redis = ">=5.0"
sentence-transformers = ">=2.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
